requirements.txt
asn1crypto ==0.24.0
certifi ==2018.11.29
cffi ==1.12.2
chardet ==3.0.4
configparser ==3.7.4
coreapi ==2.3.3
coreschema ==0.0.4
crontab ==0.22.5
cryptography ==2.6.1
defusedxml ==0.5.0
django ==2.1.7
django-auth-ldap ==1.7.0
django-cors-headers ==2.5.0
django-crontab ==0.7.1
django-rest-swagger ==2.2.0
djangorestframework ==3.9.2
docx ==0.2.4
idna ==2.8
itypes ==1.1.0
jenkinsapi ==0.3.9
jinja2 ==2.10
jira ==2.0.0
jpush ==3.3.6
kiwisolver ==1.0.1
lxml ==4.3.2
markupsafe ==1.1.1
matplotlib ==3.2.1
numpy ==1.16.2
oauthlib ==3.0.1
openapi-codec ==1.3.2
pandas ==0.24.1
pbr ==5.1.3
pyasn1 ==0.4.5
pyasn1-modules ==0.2.4
pycparser ==2.19
pymysql ==0.9.3
pyparsing ==2.3.1
python-dateutil ==2.8.0
python-docx ==0.8.10
python-ldap ==3.2.0
pytz ==2018.9
requests ==2.21.0
requests-oauthlib ==1.2.0
requests-toolbelt ==0.9.1
schedule ==0.6.0
simplejson ==3.16.0
six ==1.12.0
uritemplate ==3.0.0
urllib3 ==1.24.1
xlrd ==1.2.0
xlsxwriter ==1.1.5
xlwt ==1.3.0
Cycler ==0.10.0
Pillow ==5.4.1
PyJWT ==1.7.1